requirements.txt
beautifulsoup4==4.11.1
boto3==1.22.10
botocore==1.25.10
bs4==0.0.1
Django==4.0.4
django-crispy-forms==1.14.0
django-environ==0.8.1
django-storages==1.12.3
django-xhtml2pdf==0.0.4
lxml==4.8.0
numpy==1.22.3
opencv-python==4.5.5.64
pdf2docx==0.5.3
pdf2pptx==1.0.3
Pillow==9.0.1
psycopg2==2.9.3
python-magic==0.4.15
# python-magic-bin==0.4.14
python-pptx==0.6.21
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
s3transfer==0.5.2
selenium==4.1.3
six==1.16.0
stripe==2.73.0
urllib3==1.26.9
whitenoise==6.0.0
xhtml2pdf==0.2.7
XlsxWriter==3.0.3
apscheduler==3.9.1
PyPDF2==2.8.1
celery==5.2.7
redis==4.3.4
django-celery-results==2.4.0